# docs/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "ThreatLens AI API",
    "version": "1.0.0"
  },
  "paths": {
...
    "/api/v1/files/limits": {
      "get": {
        "tags": [
          "files"
        ],
        "summary": "Upload Limits",
        "description": "Current upload cap, so the UI can validate before sending (the cap is admin-editable).",
        "operationId": "upload_limits_api_v1_files_limits_get",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "additionalProperties": {
                    "type": "integer"
                  },
                  "type": "object",
                  "title": "Response Upload Limits Api V1 Files Limits Get"
                }
              }
            }
          }
        },
        "security": [
          {
            "OAuth2PasswordBearer": []
          }
        ]
      }
    },
    "/api/v1/files/upload": {
      "post": {
        "tags": [
          "files"
        ],
        "summary": "Upload And Scan",
        "description": "Upload a suspicious file, store it, and run the full detection pipeline.\n\nStatic analysis, ML inference, verdict fusion, detection logging and alerting\nall run (Milestone 2). Nothing is executed.",
        "operationId": "upload_and_scan_api_v1_files_upload_post",
        "requestBody": {
          "content": {
            "multipart/form-data": {
              "schema": {
                "$ref": "#/components/schemas/Body_upload_and_scan_api_v1_files_upload_post"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AnalysisResult-Output"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        },
        "security": [
          {
            "OAuth2PasswordBearer": []
          }
        ]
      }
    },
    "/api/v1/files/scan": {
      "post": {
        "tags": [
          "files"
        ],
        "summary": "Scan Stored File",
        "description": "Re-run the full pipeline on an already-stored sample (by object path).\n\nFalls back to analysing the path string itself when no stored object exists,\nso the endpoint stays usable in tests and quick demos.",
        "operationId": "scan_stored_file_api_v1_files_scan_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/AnalysisRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AnalysisResult-Output"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        },
        "security": [
          {
            "OAuth2PasswordBearer": []
          }
        ]
      }
    },
    "/api/v1/analysis/limits": {
      "get": {
        "tags": [
          "analysis"
        ],
        "summary": "Upload Limits",
        "description": "Current upload cap, so the UI can validate before sending (the cap is admin-editable).",
        "operationId": "upload_limits_api_v1_analysis_limits_get",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "additionalProperties": {
                    "type": "integer"
                  },
                  "type": "object",
                  "title": "Response Upload Limits Api V1 Analysis Limits Get"
                }
              }
            }
          }
        },
        "security": [
          {
            "OAuth2PasswordBearer": []
          }
        ]
      }
    },
    "/api/v1/analysis/upload": {
      "post": {
        "tags": [
          "analysis"
        ],
        "summary": "Upload And Scan",
        "description": "Upload a suspicious file, store it, and run the full detection pipeline.\n\nStatic analysis, ML inference, verdict fusion, detection logging and alerting\nall run (Milestone 2). Nothing is executed.",
        "operationId": "upload_and_scan_api_v1_analysis_upload_post",
        "requestBody": {
          "content": {
            "multipart/form-data": {
              "schema": {
                "$ref": "#/components/schemas/Body_upload_and_scan_api_v1_analysis_upload_post"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AnalysisResult-Output"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        },
        "security": [
          {
            "OAuth2PasswordBearer": []
          }
        ]
      }
    },
    "/api/v1/analysis/scan": {
      "post": {
        "tags": [
          "analysis"
        ],
        "summary": "Scan Stored File",
        "description": "Re-run the full pipeline on an already-stored sample (by object path).\n\nFalls back to analysing the path string itself when no stored object exists,\nso the endpoint stays usable in tests and quick demos.",
        "operationId": "scan_stored_file_api_v1_analysis_scan_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/AnalysisRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AnalysisResult-Output"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        },
        "security": [
          {
            "OAuth2PasswordBearer": []
          }
        ]
      }
    },
...
    "/api/v1/behavior/catalog": {
      "get": {
        "tags": [
          "behavior"
        ],
        "summary": "Get Catalog",
        "description": "Return the full catalog of behaviors this system can infer (static reference).",
        "operationId": "get_catalog_api_v1_behavior_catalog_get",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "additionalProperties": true,
                  "type": "object",
                  "title": "Response Get Catalog Api V1 Behavior Catalog Get"
                }
              }
            }
          }
        },
        "security": [
          {
            "OAuth2PasswordBearer": []
          }
        ]
      }
    },
    "/api/v1/behavior/tactics": {
      "get": {
        "tags": [
          "behavior"
        ],
        "summary": "Get Tactics",
        "operationId": "get_tactics_api_v1_behavior_tactics_get",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "additionalProperties": true,
                  "type": "object",
                  "title": "Response Get Tactics Api V1 Behavior Tactics Get"
                }
              }
            }
          }
        },
        "security": [
          {
            "OAuth2PasswordBearer": []
          }
        ]
      }
    },
    "/api/v1/behavior/analyze": {
      "post": {
        "tags": [
          "behavior"
        ],
        "summary": "Analyze Upload",
        "description": "Static behavioral analysis directly on an uploaded file (no execution).",
        "operationId": "analyze_upload_api_v1_behavior_analyze_post",
        "requestBody": {
          "content": {
            "multipart/form-data": {
              "schema": {
                "$ref": "#/components/schemas/Body_analyze_upload_api_v1_behavior_analyze_post"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/BehavioralAnalysisResult-Output"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        },
        "security": [
          {
            "OAuth2PasswordBearer": []
          }
        ]
      }
    },
    "/api/v1/behavior/from-analysis": {
      "post": {
        "tags": [
          "behavior"
        ],
        "summary": "Analyze From Analysis",
        "description": "Derive behavioral analysis from an existing AnalysisResult (e.g. from /files/upload).",
        "operationId": "analyze_from_analysis_api_v1_behavior_from_analysis_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/AnalysisResult-Input"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/BehavioralAnalysisResult-Output"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        },
        "security": [
          {
            "OAuth2PasswordBearer": []
          }
        ]
      }
    },
    "/api/v1/behavior/from-storage": {
      "post": {
        "tags": [
          "behavior"
        ],
        "summary": "Analyze Stored",
        "description": "Behavioral analysis for a previously uploaded object path.",
        "operationId": "analyze_stored_api_v1_behavior_from_storage_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/BehavioralAnalysisRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/BehavioralAnalysisResult-Output"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        },
        "security": [
          {
            "OAuth2PasswordBearer": []
          }
        ]
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "AnalysisRequest": {
        "properties": {
          "object_path": {
            "type": "string",
            "title": "Object Path"
          }
        },
        "type": "object",
        "required": [
          "object_path"
        ],
        "title": "AnalysisRequest"
      },
      "AnalysisResult-Input": {
        "properties": {
          "object_path": {
            "type": "string",
            "title": "Object Path"
          },
          "sha256": {
            "type": "string",
            "title": "Sha256"
          },
          "md5": {
            "type": "string",
            "title": "Md5"
          },
          "hashes": {
            "$ref": "#/components/schemas/FileHashes"
          },
          "metadata": {
            "$ref": "#/components/schemas/FileMetadata"
          },
          "signature_match": {
            "$ref": "#/components/schemas/SignatureMatch"
          },
          "yara_matches": {
            "items": {
              "$ref": "#/components/schemas/YaraMatch"
            },
            "type": "array",
            "title": "Yara Matches",
            "default": []
          },
          "yara_available": {
            "type": "boolean",
            "title": "Yara Available",
            "default": true
          },
          "network_indicators": {
            "$ref": "#/components/schemas/NetworkIndicators",
            "default": {
              "urls": [],
              "ips": [],
              "domains": []
            }
          },
          "suspicious_indicators": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Suspicious Indicators",
            "default": []
          },
          "suspicious_strings": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Suspicious Strings",
            "default": []
          },
          "strings_sample": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Strings Sample",
            "default": []
          },
          "risk": {
            "$ref": "#/components/schemas/RiskAssessment"
          },
          "behavior": {
            "$ref": "#/components/schemas/BehaviorProfile",
            "default": {
              "available": false,
              "is_pe": false,
              "capabilities": [],
              "narrative": "Behavioral analysis not run.",
              "pe_summary": {},
              "import_summary": {}
            }
          },
          "ml": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/MLPrediction"
              },
              {
                "type": "null"
              }
            ]
          },
          "verdict": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/Verdict"
              },
              {
                "type": "null"
              }
            ]
          },
          "threat_intel": {
            "$ref": "#/components/schemas/ThreatIntelResult",
            "default": {
              "configured": false,
              "available": false
            }
          },
          "behavioral_analysis": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/BehavioralAnalysisResult-Input"
              },
              {
                "type": "null"
              }
            ]
          },
          "notes": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Notes"
          }
        },
        "type": "object",
        "required": [
          "object_path",
          "sha256",
          "md5",
          "hashes",
          "metadata",
          "signature_match",
          "risk",
          "notes"
        ],
        "title": "AnalysisResult"
      },
      "AnalysisResult-Output": {
        "properties": {
          "object_path": {
            "type": "string",
            "title": "Object Path"
          },
          "sha256": {
            "type": "string",
            "title": "Sha256"
          },
          "md5": {
            "type": "string",
            "title": "Md5"
          },
          "hashes": {
            "$ref": "#/components/schemas/FileHashes"
          },
          "metadata": {
            "$ref": "#/components/schemas/FileMetadata"
          },
          "signature_match": {
            "$ref": "#/components/schemas/SignatureMatch"
          },
          "yara_matches": {
            "items": {
              "$ref": "#/components/schemas/YaraMatch"
            },
            "type": "array",
            "title": "Yara Matches",
            "default": []
          },
          "yara_available": {
            "type": "boolean",
            "title": "Yara Available",
            "default": true
          },
          "network_indicators": {
            "$ref": "#/components/schemas/NetworkIndicators",
            "default": {
              "urls": [],
              "ips": [],
              "domains": []
            }
          },
          "suspicious_indicators": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Suspicious Indicators",
            "default": []
          },
          "suspicious_strings": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Suspicious Strings",
            "default": []
          },
          "strings_sample": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Strings Sample",
            "default": []
          },
          "risk": {
            "$ref": "#/components/schemas/RiskAssessment"
          },
          "behavior": {
            "$ref": "#/components/schemas/BehaviorProfile",
            "default": {
              "available": false,
              "is_pe": false,
              "capabilities": [],
              "narrative": "Behavioral analysis not run.",
              "pe_summary": {},
              "import_summary": {}
            }
          },
          "ml": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/MLPrediction"
              },
              {
                "type": "null"
              }
            ]
          },
          "verdict": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/Verdict"
              },
              {
                "type": "null"
              }
            ]
          },
          "threat_intel": {
            "$ref": "#/components/schemas/ThreatIntelResult",
            "default": {
              "configured": false,
              "available": false
            }
          },
          "behavioral_analysis": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/BehavioralAnalysisResult-Output"
              },
              {
                "type": "null"
              }
            ]
          },
          "notes": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Notes"
          }
        },
        "type": "object",
        "required": [
          "object_path",
          "sha256",
          "md5",
          "hashes",
          "metadata",
          "signature_match",
          "risk",
          "notes"
        ],
        "title": "AnalysisResult"
      },
...
      "BehavioralAnalysisRequest": {
        "properties": {
          "object_path": {
            "type": "string",
            "title": "Object Path",
            "description": "Stored object path to re-analyse"
          }
        },
        "type": "object",
        "required": [
          "object_path"
        ],
        "title": "BehavioralAnalysisRequest"
      },
...
      "BehavioralAnalysisResult-Output": {
        "properties": {
          "generated_at": {
            "type": "string",
            "format": "date-time",
            "title": "Generated At"
          },
          "file_hash": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "File Hash"
          },
          "object_path": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Object Path"
          },
          "risk_score": {
            "type": "integer",
            "maximum": 100.0,
            "minimum": 0.0,
            "title": "Risk Score",
            "description": "Aggregate behavioral risk 0-100"
          },
          "risk_level": {
            "type": "string",
            "enum": [
              "low",
              "medium",
              "high"
            ],
            "title": "Risk Level"
          },
          "confidence": {
            "type": "number",
            "maximum": 1.0,
            "minimum": 0.0,
            "title": "Confidence",
            "description": "Mean confidence across detected behaviors"
          },
          "behaviors_detected": {
            "type": "integer",
            "title": "Behaviors Detected"
          },
          "behaviors_total": {
            "type": "integer",
            "title": "Behaviors Total"
          },
          "behaviors": {
            "items": {
              "$ref": "#/components/schemas/BehavioralFinding"
            },
            "type": "array",
            "title": "Behaviors"
          },
          "tactics_summary": {
            "items": {
              "$ref": "#/components/schemas/TacticSummary"
            },
            "type": "array",
            "title": "Tactics Summary"
          },
          "technique_coverage": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Technique Coverage",
            "description": "Unique ATT&CK technique IDs observed"
          },
          "attack_chain": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Attack Chain",
            "description": "Ordered tactics that were hit (kill-chain order)"
          },
          "summary": {
            "type": "string",
            "title": "Summary",
            "description": "Human-readable summary"
          },
          "kill_chain_stage": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Kill Chain Stage",
            "description": "Furthest kill-chain stage reached"
          }
        },
        "type": "object",
        "required": [
          "risk_score",
          "risk_level",
          "confidence",
          "behaviors_detected",
          "behaviors_total",
          "behaviors",
          "tactics_summary",
          "technique_coverage",
          "attack_chain",
          "summary"
        ],
        "title": "BehavioralAnalysisResult"
      },
...
      "Body_analyze_upload_api_v1_behavior_analyze_post": {
        "properties": {
          "file": {
            "type": "string",
            "format": "binary",
            "title": "File"
          }
        },
        "type": "object",
        "required": [
          "file"
        ],
        "title": "Body_analyze_upload_api_v1_behavior_analyze_post"
      },
      "Body_upload_and_scan_api_v1_analysis_upload_post": {
        "properties": {
          "file": {
            "type": "string",
            "format": "binary",
            "title": "File"
          }
        },
        "type": "object",
        "required": [
          "file"
        ],
        "title": "Body_upload_and_scan_api_v1_analysis_upload_post"
      },
      "Body_upload_and_scan_api_v1_files_upload_post": {
        "properties": {
          "file": {
            "type": "string",
            "format": "binary",
            "title": "File"
          }
        },
        "type": "object",
        "required": [
          "file"
        ],
        "title": "Body_upload_and_scan_api_v1_files_upload_post"
      },
...
      "FileHashes": {
        "properties": {
          "md5": {
            "type": "string",
            "title": "Md5"
          },
          "sha1": {
            "type": "string",
            "title": "Sha1"
          },
          "sha256": {
            "type": "string",
            "title": "Sha256"
          }
        },
        "type": "object",
        "required": [
          "md5",
          "sha1",
          "sha256"
        ],
        "title": "FileHashes",
        "description": "Cryptographic digests of the uploaded sample."
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
    },
    "securitySchemes": {
      "OAuth2PasswordBearer": {
        "type": "oauth2",
        "flows": {
          "password": {
            "scopes": {},
            "tokenUrl": "/api/v1/auth/login"
          }
        }
      }
    }
  }
}